add doover 2.0 APIs for CLI in a backward-compatible manner
